pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-wa-training"
version = "0.1.0"
description = "Low-resource training and retrieval pipeline for tiny-dataset bots (PDF/JSON ingestion, chunking, embeddings, FAISS index, FastAPI)."
authors = [{ name = "Kev", email = "dev@ktg.one" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "faiss-cpu>=1.8.0.post1",
  "sentence-transformers>=2.7.0",
  "transformers>=4.41.0",
  "torch>=2.3.0; platform_system != 'Windows' or platform_machine != 'ARM64'",
  "PyPDF2>=3.0.1",
  "beautifulsoup4>=4.12.3",
  "rich>=13.7.1",
  "orjson>=3.10.6",
  "typer[all]>=0.12.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.8",
  "mypy>=1.10.0",
]

[project.scripts]
aiwa = "training.cli:app"

[tool.setuptools]
packages = ["training"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]